@@ -7,14 +7,15 @@
         "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "doctrine/dbal": "^3.10",
-        "doctrine/doctrine-bundle": "^2.15",
+        "doctrine/dbal": "^3.10.1",
+        "doctrine/doctrine-bundle": "^2.15.1",
         "doctrine/doctrine-migrations-bundle": "^3.4.2",
-        "doctrine/mongodb-odm-bundle": "*",
-        "doctrine/orm": "^3.5",
-        "knpuniversity/oauth2-client-bundle": ">=2.18.3",
+        "doctrine/mongodb-odm": "^2.11.3",
+        "doctrine/mongodb-odm-bundle": "^5.3.1",
+        "doctrine/orm": "^3.5.2",
+        "knpuniversity/oauth2-client-bundle": ">=2.18.4",
         "league/oauth2-google": "^4.0.1",
-        "phpdocumentor/reflection-docblock": "^5.6.2",
+        "phpdocumentor/reflection-docblock": "^5.6.3",
         "phpstan/phpdoc-parser": "^2.2",
         "stof/doctrine-extensions-bundle": "^1.14",
         "symfony/asset": "7.3.*",
@@ -23,7 +24,7 @@
         "symfony/doctrine-messenger": "7.3.*",
         "symfony/dotenv": "7.3.*",
         "symfony/expression-language": "7.3.*",
-        "symfony/flex": "^2.8.1",
+        "symfony/flex": "^2.8.2",
         "symfony/form": "7.3.*",
         "symfony/framework-bundle": "7.3.*",
         "symfony/google-mailer": "7.3.*",
@@ -39,14 +40,14 @@
         "symfony/runtime": "7.3.*",
         "symfony/security-bundle": "7.3.*",
         "symfony/serializer": "7.3.*",
-        "symfony/stimulus-bundle": "^2.28.2",
+        "symfony/stimulus-bundle": "^2.30.0",
         "symfony/string": "7.3.*",
         "symfony/translation": "7.3.*",
         "symfony/twig-bundle": "7.3.*",
-        "symfony/ux-turbo": "^2.28.2",
+        "symfony/ux-turbo": "^2.30.0",
         "symfony/validator": "7.3.*",
         "symfony/web-link": "7.3.*",
-        "symfony/webpack-encore-bundle": "^2.2",
+        "symfony/webpack-encore-bundle": "^2.3",
         "symfony/yaml": "7.3.*",
         "twig/extra-bundle": "^2.12|^3.21",
         "twig/twig": "^2.12|^3.21.1",
@@ -104,7 +105,7 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^12.3.0",
+        "phpunit/phpunit": "^12.3.7",
         "symfony/browser-kit": "7.3.*",
         "symfony/css-selector": "7.3.*",
         "symfony/debug-bundle": "7.3.*",
